+++ src/app/middlewares/factory/withSignUpHandle.ts
@@ -5,51 +5,14 @@ import {
   NextResponse,
 } from 'next/server';
 
-import { TokenResponse } from '@api/schema/token';
-
 import {
   deleteCookie,
-  getCookie,
   setCookie,
 } from 'src/app/(route)/api/auth/[...auth]/utils/cookie';
-import {
-  decrypt,
-  encrypt,
-} from 'src/app/(route)/api/auth/[...auth]/utils/crypto';
+import { encrypt } from 'src/app/(route)/api/auth/[...auth]/utils/crypto';
 import tokenPrefix from 'src/app/(route)/api/auth/[...auth]/utils/token-prefix';
 
-const refreshUserToken = async (request: NextRequest, secret: string) => {
-  const {
-    [tokenPrefix('accessToken')]: access,
-    [tokenPrefix('refreshToken')]: refresh,
-  } = getCookie(
-    [tokenPrefix('accessToken'), tokenPrefix('refreshToken')],
-    request,
-  );
-
-  if (!access || !refresh) {
-    throw new Error('토큰이 없습니다.');
-  }
-
-  const oldAccessToken = decrypt(access, secret);
-  const oldRefreshToken = decrypt(refresh, secret);
-
-  const { accessToken, refreshToken, code, username } = await fetch(
-    `${process.env.NEXT_PUBLIC_API_DOMAIN_NAME}/api/v1/members/token/refresh`,
-    {
-      method: 'POST',
-      headers: {
-        Authorization: `Bearer ${oldAccessToken}`,
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        refreshToken: oldRefreshToken,
-      }),
-    },
-  ).then((res) => res.json() as Promise<TokenResponse>);
-
-  return { accessToken, refreshToken, code, username };
-};
+import refreshUserToken from '../utils/refresh-token';
 
 function withSignUpHandle(middleware: NextMiddleware) {
   return async (request: NextRequest, event: NextFetchEvent) => {
